Conflicts:
	Sampoerna.EMS.AutoMapperExtensions/CustomResolvers.cs
	Sampoerna.EMS.BLL/MasterDataBLL.cs
	Sampoerna.EMS.BLL/PlantBLL.cs
	Sampoerna.EMS.Contract/IMasterDataBLL.cs
	Sampoerna.EMS.Contract/IZaidmExNPPBKCBLL.cs
	Sampoerna.EMS.Website/EMSWebsiteMapper.cs
	Sampoerna.EMS.Website/Sampoerna.EMS.Website.csproj

+++ Sampoerna.EMS.Contract/IZaidmExProdTypeBLL.cs
@@ -5,7 +5,7 @@ namespace Sampoerna.EMS.Contract
 {
     public interface IZaidmExProdTypeBLL
     {
-        ZAIDM_EX_PRODTYP GetById(int id);
+        ZAIDM_EX_PRODTYP GetById(long id);
 
         List<ZAIDM_EX_PRODTYP> GetAll();
         ZAIDM_EX_PRODTYP GetByCode(int Code);
